Refactor Kafka Reception and Delivery beans
Split out an abstract class for easier reuse
